Move MaterialProperties and MaterialPipeline to bevy_material
